python-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
pymongo==4.15.5
dnspython==2.8.0
python-dotenv==1.0.0
